pub/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
joblib==1.5.1
    # via scikit-learn
kafka-python==2.2.15
    # via -r requirements.in
numpy==2.3.2
    # via
    #   scikit-learn
    #   scipy
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.16.1
    # via scikit-learn
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via -r requirements.in
